+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K82WV7KNHN02VYPYP6RCEH8F"
+        "01K836P72PSGVWEJVCAFAWX3KV"
     )
     r"""The id of the resource"""