+++ packages/orq-rc/src/orq_ai_sdk/models/listbudgetsop.py
@@ -242,7 +242,7 @@ class ListBudgetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-21T08:05:43.551Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-21T10:57:44.779Z")
     r"""The date and time the resource was last updated"""