+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K82WV7KKNVS9Q61G57MJDMT3"
+        "01K836P72N3P2S9JT6EW7V40NH"
     )
     r"""The id of the resource"""