+++ packages/orq-rc/RELEASES.md
@@ -3459,3 +3459,13 @@ Based on:
 - [python v3.14.0-rc.41] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.41] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.41 - packages/orq-rc
+
+## 2025-10-21 10:59:32
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.637.3 (2.727.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.42] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.42] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.42 - packages/orq-rc